requirements.txt
uvicorn
pydantic
requests
fastapi
python_dotenv